requirements.txt
requests
pytest
fastapi
uvicorn
python-dotenv
sqlalchemy
psycopg2-binary
pandas
plotly
pillow
kaleido
